Move custom connect button to own component. Adjust basic styles
